requirements.txt
streamlit
pytesseract
pillow
scikit-learn
xgboost
pandas
numpy
matplotlib
plotly
seaborn
pyzbar
